refactor: 修改HabitType定义位置
1. 将HabitType移入HabitModel
