+++ lib/model/habit_model.dart
@@ -1,5 +1,8 @@
 import 'package:drift/drift.dart';
+import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:liferpg/model/common_model.dart';
+
+enum HabitType { good, bad }
 
 @DataClassName("HabitModel")
 class HabitTable extends Table {
@@ -22,4 +25,15 @@ class HabitTable extends Table {
   DateTimeColumn get lastFinishedAt => dateTime()();
 
   DateTimeColumn get createdAt => dateTime()();
+}
+
+extension HabitTypeExtension on HabitType {
+  String localizedString(context) {
+    switch (this) {
+      case HabitType.good:
+        return AppLocalizations.of(context)!.good;
+      case HabitType.bad:
+        return AppLocalizations.of(context)!.bad;
+    }
+  }
 }
