+++ lib/view/habit_edit_view.dart
@@ -3,6 +3,7 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
 import '../database/database.dart';
 import '../model/common_model.dart';
+import '../model/habit_model.dart';
 import '../viewmodel/habit_viewmodel.dart';
 
 class HabitEditView extends StatefulWidget {
